requirements.txt
streamlit
boto3
langchain
opensearch-py
python-dotenv